+++ core/properties/__init__.py
@@ -1,0 +1,8 @@
+from core.properties.common import FluidTransportProperties
+from core.properties.fluids import ConstantPropertyProvider, PropertyProvider
+
+__all__ = [
+    "FluidTransportProperties",
+    "PropertyProvider",
+    "ConstantPropertyProvider",
+]